requirements.txt
flask==2.3.3
flask-cors==4.0.0
rembg==2.0.50
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
psutil==5.9.6